requirements.txt
pandas
pyyaml
streamlit
matplotlib
seaborn
